pyproject.toml
[project]
name = "tableplus-db-urls"
description = "Generate TablePlus DB URLs from CookieCutter Django to make setting up connections easier."
version = "0.2.0"
dynamic = ["readme"]
requires-python = ">=3.11"
authors = [
  { name = "Tim Santor", email = "tsantor@xstudios.com" },
]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Natural Language :: English",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
]
dependencies = [
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
cli = [
  "click>=8.1.8",
]

[dependency-groups]
dev=[
  "hatch>=1.16.5",
  "pre-commit>=4.4.0",
  "ruff>=0.14.4",
  "twine>=6.2.0",
  "xapp-tools>=0.4.0",
]
test=[
  "click>=8.1.8",
  "pytest-cov>=7.0.0",
  "pytest-mock>=3.15.1",
  "tox-uv>=1.29.0",
]

[project.urls]
Repository = "https://github.com/tsantor/tableplus-db-urls.git"
Issues = "https://github.com/tsantor/tableplus-db-urls/issues"
Changelog = "https://github.com/tsantor/tableplus-db-urls/blob/master/HISTORY.md"

[project.scripts]
tableplus = "tableplus.cli:cli"

[build-system]
requires = ["hatchling", "xapp-tools"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tableplus"]

[tool.hatch.metadata.hooks.combined-readme]

# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
python_files = ["tests.py", "test_*.py"]
norecursedirs = ["node_modules"]
filterwarnings = "ignore::DeprecationWarning"
pythonpath = "src"

# ==== Coverage ====
[tool.coverage.run]
omit = ["*/migrations/*", "*/tests/*"]

[tool.coverage.report]
exclude_also = [
  "def __repr__",
]

# ==== mypy ====
[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[tool.ruff]
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
  "*/migrations/*.py",
  "staticfiles/*",
]
line-length = 88
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
select = [
  "F",
  "E",
  "W",
  "C90",
  "I",
  "N",
  "UP",
  "YTT",
  "ASYNC",
  "S",
  "BLE",
  "FBT",
  "B",
  "A",
  "COM",
  "C4",
  "DTZ",
  "T10",
  "DJ",
  "EM",
  "EXE",
  "FA",
  "ISC",
  "ICN",
  "G",
  "INP",
  "PIE",
  "T20",
  "PYI",
  "PT",
  "Q",
  "RSE",
  "RET",
  "SLF",
  "SLOT",
  "SIM",
  "TID",
  "TCH",
  "INT",
  "PTH",
  "ERA",
  "PD",
  "PGH",
  "PL",
  "TRY",
  "FLY",
  "PERF",
  "RUF",
]
ignore = [
  "S101",
  "RUF012",
  "SIM102",
  "ERA001",
  "PGH004",
  "E501",
]
extend-ignore = [
  "S106",
  "RUF005",
  "S104",
  "S311",
  "FBT001",
  "FBT002",
  "S308",
  "TRY300",
  "COM812",
  "ISC001",
]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
force-single-line = true
